Make multi-class work
This commit overhauls the macro, using a two-pass parse/codegen
approach that makes it easier for us to implement some long-desired
features. This commit adds support for multiple classes in a single
macro.
